refactor: move pretty repr tests back to their own file
